lib/client/requirements.txt
lxml
pandas
typing
pydantic
fastapi
uvicorn[standard]
gunicorn
requests

# Install solathon dependencies
anyio==4.1.0
base58==2.1.1
certifi==2023.11.17
cffi==1.16.0
construct==2.10.70
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.22.0
idna==3.6
pycparser==2.21
PyNaCl==1.5.0
Pillow==10.1.0
sniffio==1.3.0
qrcode==7.4.2